Implementing a simple alternative to Collapse from MUI
