+++ src/components/AltinnCollapsable.module.css
@@ -1,0 +1,22 @@
+.collapsable {
+  overflow: hidden;
+  transition:
+    max-height 0.3s ease-in-out,
+    opacity 0.2s ease-in-out;
+  will-change: max-height;
+  box-sizing: border-box;
+}
+
+.collapsableOpen {
+  max-height: 1000px; /* Set to a value larger than your expected content */
+}
+
+.collapsableClosed {
+  max-height: 0;
+  padding-top: 0;
+  padding-bottom: 0;
+  margin-top: 0;
+  margin-bottom: 0;
+  border-top-width: 0;
+  border-bottom-width: 0;
+}
